% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\name{check_products_column_names}
\alias{check_products_column_names}
\title{Check if column names of applied pesticide products dataframe are valid}
\usage{
check_products_column_names(products)
}
\arguments{
\item{products}{Dataframe with raw pesticide application data.}
}
\description{
Check if column names of applied pesticide products dataframe are valid
}
