% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\name{check_substance_column_names}
\alias{check_substance_column_names}
\title{Check if column names of substances dataframe are valid}
\usage{
check_substance_column_names(substances)
}
\arguments{
\item{substances}{Dataframe describing active ingredients of the applied pesticide products.}
}
\description{
Check if column names of substances dataframe are valid
}
