% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\name{compute_pesticide_load_indicator_ppdb}
\alias{compute_pesticide_load_indicator_ppdb}
\title{Compute Pesticide Load Indicator using the Pesticide Properties database}
\usage{
compute_pesticide_load_indicator_ppdb(substances, products, folder)
}
\arguments{
\item{substances}{Dataframe describing active ingredients of the applied pesticide products and their CAS number.}

\item{products}{Dataframe with raw pesticide application data.}

\item{folder}{Folder with exported xlsx files from PPDB containing information on active ingredient properties.}
}
\value{
products Pesticide application dataframe with pesticide indicator columns added.
}
\description{
Compute Pesticide Load Indicator using the Pesticide Properties database
}
