% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\name{compute_pesticide_load_indicator_single}
\alias{compute_pesticide_load_indicator_single}
\title{Compute Pesticide Load Indicator for a single application using the Pesticide Properties database}
\usage{
compute_pesticide_load_indicator_single(
  folder,
  product,
  year,
  formula,
  substances,
  cas_numbers,
  concentrations,
  ...
)
}
\arguments{
\item{folder}{Folder with exported xlsx files from PPDB containing information on active ingredient properties.}

\item{product}{Product name of the applied pesticide.}

\item{year}{Application year.}

\item{formula}{Load formulation factor for applied product.}

\item{substances}{List or vector of active ingredient names of the applied pesticide.}

\item{cas_numbers}{List or vector of CAS numbers of the respective active ingredients.}

\item{concentrations}{List or vector of product concentrations of the respective active ingredients.}

\item{...}{overrides for default Load factors.}
}
\value{
products Dataframe with computed pestice Loads.
}
\description{
Compute Pesticide Load Indicator for a single application using the Pesticide Properties database
}
