% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ppdb.R
\name{compute_risk_score}
\alias{compute_risk_score}
\title{Calculates the sum of risk scores from a list of H-phrases and expands the table}
\usage{
compute_risk_score(table, var_name)
}
\arguments{
\item{table}{Dataframe with H-phrases on product level.}

\item{var_name}{Name of the variable that contains the information (string) on H-phrases.
For example "H317; H411" or "H410, H411".}
}
\value{
Adds a variable indicating the sum of risk scores
needed to compute the Pesticide Load Indicator
Check national pesticide databases for information on product labels of pesticides
(information on H-phrases of each product).
}
\description{
Calculates the sum of risk scores from a list of H-phrases and expands the table
}
