% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\docType{data}
\name{default.load.factors}
\alias{default.load.factors}
\title{Default load factors}
\format{
An object of class \code{list} of length 14.
}
\usage{
default.load.factors
}
\description{
Default load factors
}
\keyword{datasets}
