% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\name{products.path}
\alias{products.path}
\title{path to  includedexamples products.xlsx file}
\usage{
products.path()
}
\value{
path to products.xlsx file
}
\description{
path to  includedexamples products.xlsx file
}
