% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\name{substances.load}
\alias{substances.load}
\title{load included substances.xlsx file}
\usage{
substances.load()
}
\value{
substances.xlsx file as data.frame
}
\description{
load included substances.xlsx file
}
