% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\name{substances.path}
\alias{substances.path}
\title{path to included examples substances.xlsx file}
\usage{
substances.path()
}
\value{
path to substances.xlsx file
}
\description{
path to included examples substances.xlsx file
}
