% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_load_indicator.R
\name{compute_pesticide_load_indicator}
\alias{compute_pesticide_load_indicator}
\title{Compute Pesticide Load Indicator with user supplied information on pesticide properties}
\usage{
compute_pesticide_load_indicator(substances, products)
}
\arguments{
\item{substances}{Dataframe describing active ingredients of the applied pesticide products, including their ecotoxicity, fate and human health properties.}

\item{products}{Dataframe with raw pesticide application data.}
}
\value{
products Pesticide application dataframe with pesticide indicator columns added.
}
\description{
Compute Pesticide Load Indicator with user supplied information on pesticide properties
}
